refactor and redirects
